requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Pillow==8.3.2
requests==2.26.0
model-mommy==2.0.0